refactory stremaing lock
